make ids longs instead of ints, add group creation

+++ app/src/main/java/com/projectblejder/jointcash/infrastructure/models/Expense.kt
@@ -15,7 +15,7 @@ import android.arch.persistence.room.PrimaryKey
 )
 data class Expense(
         @PrimaryKey(autoGenerate = true)
-        var id: Int?,
+        var id: Long?,
 
         @ColumnInfo(name = "name")
         var name: String,
